Added README.md documentation
