+++ react/components/AutocompleteInput/Option.tsx
@@ -2,6 +2,27 @@ import classnames from 'classnames'
 import PropTypes from 'prop-types'
 import React, { useState } from 'react'
 
+/** Structured Autocomplete option shape */
+const structuredAutocompleteOptionShape = PropTypes.shape({
+  value: PropTypes.any,
+  label: PropTypes.string.isRequired,
+})
+
+type StructuredAutocompleteOption = PropTypes.InferType<
+  typeof structuredAutocompleteOptionShape
+>
+
+export type AutocompleteOption = string | StructuredAutocompleteOption
+
+/** Autocomplete option shape */
+export const autocompleteOptionShape = PropTypes.oneOfType([
+  PropTypes.string,
+  structuredAutocompleteOptionShape,
+])
+ 
+export const getTermFromOption = (option: AutocompleteOption): string =>
+  typeof option === 'string' ? option : option.label
+
 const propTypes = {
   /** Determine if the option should have a rounded bottom */
   roundedBottom: PropTypes.bool,
@@ -10,7 +31,7 @@ const propTypes = {
   /** Search term used to highlight it into the label */
   searchTerm: PropTypes.string.isRequired,
   /** Option title */
-  value: PropTypes.string.isRequired,
+  value: autocompleteOptionShape.isRequired,
   /** Option key used in the list */
   key: PropTypes.string.isRequired,
   /** Determine if an option is selected and should be highlighted */
@@ -22,16 +43,9 @@ const propTypes = {
 const Option: React.FunctionComponent<
   PropTypes.InferProps<typeof propTypes>
 > = props => {
-  const {
-    icon,
-    selected,
-    roundedBottom,
-    searchTerm,
-    value,
-    key,
-    onClick,
-  } = props
+  const { icon, selected, roundedBottom, searchTerm, key, onClick } = props
   const [highlightOption, setHighlightOption] = useState(false)
+  const value = getTermFromOption(props.value)
 
   const renderOptionValue = (): string | React.ReactElement => {
     const index = value.toLowerCase().indexOf(searchTerm.toLowerCase())
